Update layouts pages
